add ui5-toolign if missing
